Add router and pages for sport and news
