+++ package.json
@@ -10,11 +10,13 @@
     "@emotion/react": "^11.1.5",
     "@guardian/src-button": "^3.5.0",
     "react": "^17.0.0",
-    "react-dom": "^17.0.0"
+    "react-dom": "^17.0.0",
+    "react-router-dom": "^5.2.0"
   },
   "devDependencies": {
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
+    "@types/react-router-dom": "^5.1.7",
     "@vitejs/plugin-react-refresh": "^1.3.1",
     "typescript": "^4.1.2",
     "vite": "^2.1.5"
