+++ src/news/News.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const News = () => {
+  return (
+    <div>
+      News
+    </div>
+  );
+};
+
+export { News };
