+++ src/page/Page.tsx
@@ -1,32 +1,22 @@
-import React, { useContext, useEffect, useState } from "react";
-import "./Page.css";
-import Data from "../ListData.json";
-import { Card } from "../card/Card";
-import FilterBank from "../filterBank/filterBank";
-import { FilterContext } from "../filterContext/filterContext";
+import React, { FC } from 'react';
+import FilterContextWrapper from "../filterContext/filterContext";
+import './Page.css';
 
-const Page = () => {
-	const { currentFilter } = useContext(FilterContext);
-	const [list, setList] = useState(Data.cards);
+interface Props {
+	children: React.ReactNode;
+}
 
-	useEffect(() => {
-		const filteredCards = Data.cards.filter(
-			(card) => card.tag === currentFilter
-		);
-		setList(filteredCards);
-	}, [currentFilter]);
-
-	return (
-		<section className="Page-body">
-			<div className="Page-container">
-				<div className="List-container">
-					<FilterBank />
-					{list.map((card: any) => (
-						<Card image={card.image} tag={card.tag} />
-					))}
-				</div>
-			</div>
-		</section>
+const Page: FC<Props> = ({ children }) => {
+    return  (
+        <FilterContextWrapper>
+            <section className="Page-body">
+                <div className="Page-container">
+                    <div className="List-container">
+                        {children}
+                    </div>
+                </div>
+            </section>
+        </FilterContextWrapper>
 	);
 };
 export { Page };
